requirements.txt
discord==1.7.3
python-dotenv==0.21.1
openpyxl==3.1.1
pyairtable==1.4.0